Add inputrc, ackrc, git aliases. Fix bash_profile and symlink script

+++ symlink.sh
@@ -4,9 +4,13 @@ DIR=`pwd`
 
 FILES=`ls -a | grep -v symlink | egrep -v '.git$' | egrep -v '^\.+$' | grep -v README`
 
-
 for f in $FILES
 do
-  echo $f
-  ln -s $DIR/$f $HOME/$f
+  if [ -e ~/$f ]
+  then
+    echo ~/$f already exists
+  else
+    ln -s $DIR/$f $HOME/$f
+    echo symlink created from $DIR/$f to ~/$f
+  fi
 done
